Struct ws::Error

pub struct Error {
    pub kind: Kind,
    pub details: Cow<'static, str>,
}

A struct indicating the kind of error that has occured and any precise details of that error.

Fields
kind: Kind details: Cow<'static, str>
Methods

impl Error

pub fn new<I>(kind: Kind, details: I) -> Error where
I: Into<Cow<'static, str>>,

Important traits for Box<R>
impl<R> Read for Box<R> where
R: Read + ?Sized, impl<W> Write for Box<W> where
W: Write + ?Sized, impl<I> Iterator for Box<I> where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn into_box(self) -> Box<StdError>