Struct ws::Request

pub struct Request {
    // some fields omitted
}

The handshake request.

Methods

impl Request

fn origin(&self) -> Result<Option<&str>>

Get the origin of the request if it comes from a browser.

fn key(&self) -> Result<&[u8]>

Get the WebSocket key sent in the request.

fn version(&self) -> Result<&str>

Get the WebSocket protocol version from the request (should be 13).

fn resource(&self) -> Result<&str>

Get the path of the request.

fn protocols(&self) -> Result<Vec<&str>>

Get the possible protocols the other endpoint supports.

fn extensions(&self) -> Result<Vec<&str>>

Get the extensions that the other endpoint is trying to negotiate.

fn headers(&self) -> Result<Vec<(&str, &str)>>

Access the request headers as a vector of tuples.