Struct ws::Response

pub struct Response {
    // some fields omitted
}

The handshake response.

Methods

impl Response

fn status(&self) -> Result<u16>

Get the status of the response

fn key(&self) -> Result<&[u8]>

Get the hashed WebSocket key

fn protocol(&self) -> Result<Option<&str>>

Get the protocol that the server has decided to use

fn extensions(&self) -> Result<Vec<&str>>

Get the extensions that the server has decided to use. If these are unacceptable, it is appropriate to send an Extension close code