Struct ws::Sender

pub struct Sender {
    // some fields omitted
}

A representation of the output of the WebSocket connection. Use this to send messages to the other endpoint.

Methods

impl Sender

fn token(&self) -> Token

A Token identifying this sender within the WebSocket.

fn send<M>(&self, msg: M) -> Result<()> where M: Into<Message>

Send a message over the connection.

fn broadcast<M>(&self, msg: M) -> Result<()> where M: Into<Message>

Send a message to the endpoints of all connections.

Be careful with this method because it does not discriminate between client and server connections, only connections that are not part of this side of the WebSocket. If your WebSocket is only functioning as a server, then usage is simple, however if you have a WebSocket that is listening for connections and is also connected to another WebSocket, this method will broadcast a message to all the clients connected and to the server at the other end of the single client connection.

fn close(&self, code: CloseCode) -> Result<()>

Send a close code to the other endpoint.

fn close_with_reason<S>(&self, code: CloseCode, reason: S) -> Result<()> where S: Into<Cow<'static, str>>

Send a close code and provide a descriptive reason for closing.

fn ping(&self, data: Vec<u8>) -> Result<()>

Send a ping to the other endpoint with the given test data.

fn pong(&self, data: Vec<u8>) -> Result<()>

Send a pong to the other endpoint responding with the given test data.

fn connect(&self, url: Url) -> Result<()>

Queue a new connection on this WebSocket to the specified URL.

fn shutdown(&self) -> Result<()>

Request that all connections terminate and that the WebSocket stop running.

fn timeout(&self, ms: u64, token: Token) -> Result<()>

Schedule a token to be sent to the WebSocket Handler's on_timeout method after ms milliseconds

fn cancel(&self, timeout: Timeout) -> Result<()>

Queue the cancellation of a previously scheduled timeout.

This method is not guaranteed to prevent the timeout from occuring, because it is possible to call this method after a timeout has already occured. It is still necessary to handle spurious timeouts.