Struct ws::WebSocketSettings

pub struct WebSocketSettings {
    pub max_connections: usize,
    pub panic_on_new_connection: bool,
    pub panic_on_shutdown: bool,
    pub protocols: Option<&'static str>,
    pub extensions: Option<&'static str>,
}

Settings that apply to multiple connections.

Fields

max_connections

The maximum number of connections that this WebSocket will support. Default: 10,000

panic_on_new_connection

Whether to panic when unable to establish a new TCP connection. Default: false

panic_on_shutdown

Whether to panic when a shutdown of the WebSocket is requested. Default: false

protocols

A protocol string representing the subprotocols that this WebSocket can support. This will be sent in Requests to server endpoints to help determine a subprotocol if any for the connection. Default: None

extensions

A WebSocket extension string indicating the extensions that this WebSocket can support. Default: None