Trait ws::Handler

pub trait Handler {
    fn settings(&mut self) -> Settings { ... }
    fn on_shutdown(&mut self) { ... }
    fn on_open(&mut self, shake: Handshake) -> Result<()> { ... }
    fn on_message(&mut self, msg: Message) -> Result<()> { ... }
    fn on_close(&mut self, code: CloseCode, reason: &str) { ... }
    fn on_error(&mut self, err: Error) { ... }
    fn on_request(&mut self, req: &Request) -> Result<(Option<&str>, Option<&str>)> { ... }
    fn on_response(&mut self, res: &Response) -> Result<()> { ... }
    fn on_ping_frame(&mut self, frame: Frame) -> Result<Option<Frame>> { ... }
    fn on_pong_frame(&mut self, frame: Frame) -> Result<Option<Frame>> { ... }
    fn on_close_frame(&mut self, frame: Frame) -> Result<Option<Frame>> { ... }
    fn on_binary_frame(&mut self, frame: Frame) -> Result<Option<Frame>> { ... }
    fn on_text_frame(&mut self, frame: Frame) -> Result<Option<Frame>> { ... }
    fn on_fragmented_frame(&mut self, frame: Frame) -> Result<Option<Frame>> { ... }
}

The core trait of this library. Implementing this trait provides the business logic of the WebSocket application.

Provided Methods

fn settings(&mut self) -> Settings

fn on_shutdown(&mut self)

Called when a request to shutdown all connections has been received.

fn on_open(&mut self, shake: Handshake) -> Result<()>

Called when the WebSocket handshake is successful and the connection is open for sending and receiving messages.

fn on_message(&mut self, msg: Message) -> Result<()>

Called on incoming messages.

fn on_close(&mut self, code: CloseCode, reason: &str)

Called when the other endpoint is asking to close the connection.

fn on_error(&mut self, err: Error)

Called when an error occurs on the WebSocket.

fn on_request(&mut self, req: &Request) -> Result<(Option<&str>, Option<&str>)>

A method for handling the low-level workings of the request portion of the WebSocket handshake. Implementors can inspect the Request and select an appropriate string protocol and/or extensions to be used. This method will not be called when the handler represents a client endpoint.

fn on_response(&mut self, res: &Response) -> Result<()>

A method for handling the low-level workings of the response portion of the WebSocket handshake. Implementors can inspect the Response and choose to faile the connection by returning an error. This method will not be called when the handler represents a server endpoint.

fn on_ping_frame(&mut self, frame: Frame) -> Result<Option<Frame>>

fn on_pong_frame(&mut self, frame: Frame) -> Result<Option<Frame>>

fn on_close_frame(&mut self, frame: Frame) -> Result<Option<Frame>>

fn on_binary_frame(&mut self, frame: Frame) -> Result<Option<Frame>>

fn on_text_frame(&mut self, frame: Frame) -> Result<Option<Frame>>

fn on_fragmented_frame(&mut self, frame: Frame) -> Result<Option<Frame>>