Enum ws::Message

pub enum Message {
    Text(String),
    Binary(Vec<u8>),
}

An enum representing the various forms of a WebSocket message.

Variants

Text

A text WebSocket message

Binary

A binary WebSocket message

Methods

impl Message

fn text<S>(string: S) -> Message where S: Into<String>

Create a new text WebSocket message from a stringable.

fn binary<B>(bin: B) -> Message where B: Into<Vec<u8>>

Create a new binary WebSocket message by converting to Vec.

fn is_text(&self) -> bool

Indicates whether a message is a text message.

fn is_binary(&self) -> bool

Indicates whether a message is a binary message.

fn len(&self) -> usize

Get the length of the WebSocket message.

fn is_empty(&self) -> bool

Returns true if the WebSocket message has no content. For example, if the other side of the connection sent an empty string.

fn into_data(self) -> Vec<u8>

Consume the WebSocket and return it as binary data.

fn into_text(self) -> Result<String>

Attempt to consume the WebSocket message and convert it to a String.

fn as_text(&self) -> Result<&str>

Attempt to get a &str from the WebSocket message, this will try to convert binary data to utf8.