Struct ws::Error

pub struct Error {
    pub kind: Kind,
    pub details: Cow<'static, str>,
}

A struct indicating the kind of error that has occured and any precise details of that error.

Fields

kind

details

Methods

impl Error

fn new<I>(kind: Kind, details: I) -> Error where I: Into<Cow<'static, str>>

fn into_box(self) -> Box<StdError>