Struct ws::Response

pub struct Response { /* fields omitted */ }

The handshake response.

Methods

impl Response

pub fn new<R>(status: u16, reason: R, body: Vec<u8>) -> Response where
R: Into<String>,

Construct a generic HTTP response with a body.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn body(&self) -> &[u8]

Get the response body.

pub fn header_mut(&mut self, header: &str) -> Option<&mut Vec<u8>>

Edit the value of the first instance of an HTTP header.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn headers(&self) -> &Vec<(String, Vec<u8>)>

Access the request headers.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn headers_mut(&mut self) -> &mut Vec<(String, Vec<u8>)>

Edit the request headers.

pub fn status(&self) -> u16

Get the HTTP status code.

pub fn set_status(&mut self, status: u16)

Set the HTTP status code.

pub fn reason(&self) -> &str

Get the HTTP status reason.

pub fn set_reason<R>(&mut self, reason: R) where
R: Into<String>,

Set the HTTP status reason.

pub fn key(&self) -> Result<&Vec<u8>>

Get the hashed WebSocket key.

pub fn protocol(&self) -> Result<Option<&str>>

Get the protocol that the server has decided to use.

pub fn set_protocol(&mut self, protocol: &str)

Set the protocol that the server has decided to use.

pub fn extensions(&self) -> Result<Vec<&str>>

Get the extensions that the server has decided to use. If these are unacceptable, it is appropriate to send an Extension close code.

pub fn add_extension(&mut self, ext: &str)

Add an accepted extension to this response. This may result in duplicate extensions listed.

pub fn remove_extension(&mut self, ext: &str)

Remove an accepted extension from this response. This will remove all configurations of the extension.

pub fn parse(buf: &[u8]) -> Result<Option<Response>>

Attempt to parse an HTTP response from a buffer. If the buffer does not contain a complete response, thiw will return Ok(None).

pub fn from_request(req: &Request) -> Result<Response>

Construct a new WebSocket handshake HTTP response from a request. This will create a response that ignores protocols and extensions. Edit this response to accept a protocol and extensions as necessary.

pub fn format<W>(&self, w: &mut W) -> Result<()> where
W: Write,

Write a response out to a buffer